% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trajectory.R
\name{calculate_angles}
\alias{calculate_angles}
\title{Calculate Trajectory Turn Angles}
\usage{
calculate_angles(trajectory, bouts = NULL)
}
\arguments{
\item{trajectory}{a trajectory data frame. See `create_trajectory()` for more
details.}

\item{bouts}{a vector of indices corresponding to beginnings of bouts. If
this argument is provided, this function only uses observations at bout
starts to calculate turn angles. Hence, turn angles are calculated using
every consecutive group of three bouts. See `find_bouts()` for more
details.}
}
\value{
a vector of turn angles. Each turn angle has a sign indicating its
  direction: left turns are positive while right turns are negative. In the
  edge case where a zebrafish somehow turns exactly 180 degrees, the sign is
  positive regardless of whether a left or right turn was performed.
}
\description{
This function calculates turn angles formed by every consecutive
  group of three observations within the trajectory data frame.
}
\examples{
trajectory <- create_trajectory(cadmium_zebrafish,
                                x = "X.center",
                                y = "Y.center",
                                t = "Trial.time")
bouts <- find_bouts(trajectory, min_distance = 2, min_time = 0.6)
angles <- calculate_angles(trajectory, bouts = bouts)

}
\references{
Tadres, D., & Louis, M. (2020). PiVR: An affordable and versatile
  closed-loop platform to study unrestrained sensorimotor behavior.
  \emph{PLoS Biology}, \emph{18}(7), e3000712.
  https://doi.org/10.1371/journal.pbio.3000712
}
