% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trajectory.R
\name{calculate_bouts_per_time_unit}
\alias{calculate_bouts_per_time_unit}
\title{Calculate Trajectory Bouts per a Specified Unit of Time}
\usage{
calculate_bouts_per_time_unit(trajectory, bouts, unit = 1)
}
\arguments{
\item{trajectory}{a trajectory data frame. See `create_trajectory()` for more
details.}

\item{bouts}{a vector of indices corresponding to beginnings of bouts. See
`find_bouts()` for more details.}

\item{unit}{the unit of time to group bouts by. This unit must be the same as
the unit used for times within the trajectory data frame. For example, if
trajectory times were recorded in seconds, setting this argument to 60
would calculate the number of bouts per minute.}
}
\value{
a vector of bouts per time unit.

  Note that the last element of this vector may be incomplete. For example,
  if time were recorded in seconds and the `unit` argument was set to 60,
  then the first element would indicate the number of bouts occurring at time
  greater than or equal to 0 seconds and less than 60 seconds, the second
  element would indicate the number of bouts occurring at time greater than
  or equal to 60 seconds and less than 120 seconds, and so on.

  If only 90 seconds were recorded in the trajectory data frame, the return
  value would have two elements. The first element would indicate the number
  of bouts occurring at time greater than or equal to 0 seconds and less than
  60 seconds, and the second element would indicate the number of bouts
  occurring at time greater than or equal to 60 seconds and less than or
  equal to 90 seconds, despite how 60 seconds was specified as the time unit.

  If exactly 60 seconds were recorded in the trajectory data frame, the
  return value would have two elements. The first element would indicate the
  number of bouts occurring at time greater than or equal to 0 seconds and
  less than 60 seconds, and the second element would indicate the number of
  bouts occurring at the exact 60-second mark (which logically should not be
  greater than 1).
}
\description{
This function groups bouts using a time unit and calculates the
  number of bouts within each unit.
}
\examples{
trajectory <- create_trajectory(cadmium_zebrafish,
                                x = "X.center",
                                y = "Y.center",
                                t = "Trial.time")
bouts <- find_bouts(trajectory, min_distance = 2, min_time = 0.6)
bouts_per_minute <- calculate_bouts_per_time_unit(trajectory,
                                                  bouts,
                                                  unit = 60)

}
