% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trajectory.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculate Trajectory Distances}
\usage{
calculate_distances(trajectory, bouts = NULL)
}
\arguments{
\item{trajectory}{a trajectory data frame. See `create_trajectory()` for more
details.}

\item{bouts}{a vector of indices corresponding to beginnings of bouts. If
this argument is provided, this function only uses observations at bout
starts to calculate distances. Hence, distances traveled are calculated
between every consecutive pair of bouts. See `find_bouts()` for more
details.}
}
\value{
a vector of distances traveled.
}
\description{
This function calculates distances traveled between every
  consecutive pair of observations within the trajectory data frame.
}
\examples{
trajectory <- create_trajectory(cadmium_zebrafish,
                                x = "X.center",
                                y = "Y.center",
                                t = "Trial.time")
bouts <- find_bouts(trajectory, min_distance = 2, min_time = 0.6)
distances <- calculate_distances(trajectory, bouts = bouts)

}
