% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trajectory.R
\name{create_trajectory}
\alias{create_trajectory}
\title{Create a Trajectory Data Frame}
\usage{
create_trajectory(data, x, y, t = NULL, rate = NULL)
}
\arguments{
\item{data}{a data frame containing the x- and y-positions of a fish.}

\item{x}{the name of the column containing the x-positions of the fish.}

\item{y}{the name of the column containing the y-positions of the fish.}

\item{t}{if a time column is present in the data, this parameter is the name
of the column containing the times of all observations.}

\item{rate}{an alternative way of specifying time. This parameter generates a
sequence of numbers starting from 0 and changing by `rate` for each
observation.}
}
\value{
a data frame containing the trajectory of a zebrafish that can be
  used by analysis functions within this package.
}
\description{
This function produces a valid data frame able to be used by
  other plotting and analysis functions in this package that contains the
  trajectory of a zebrafish. The trajectory data frame has three columns:
  x-position, y-position, and time.
}
\examples{
trajectory <- create_trajectory(cadmium_zebrafish,
                                x = "X.center",
                                y = "Y.center",
                                t = "Trial.time")

}
