% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{linear_algebra_angles}
\alias{linear_algebra_angles}
\title{Calculate Directed Angles}
\usage{
linear_algebra_angles(v1, v2)
}
\arguments{
\item{v1}{a matrix with n rows and 2 columns specifying n 2D vectors that
each are an initial side of a directed angle.}

\item{v2}{a matrix with n rows and 2 columns specifying n 2D vectors that
each are a terminal side of a directed angle.}
}
\value{
A vector containing the magnitude and direction of each directed
  angle. A positive sign indicates a counterclockwise angle, while a negative
  sign indicates a clockwise angle.
}
\description{
This function uses linear algebra and vectorized operations to
  quickly calculate many directed angles formed by pairs of vectors. Vectors
  are assumed to have their tails at the origin of a 2D coordinate plane.
}
