% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_algebra.R
\name{linear_algebra_distances}
\alias{linear_algebra_distances}
\title{Calculate Distances}
\usage{
linear_algebra_distances(p1, p2)
}
\arguments{
\item{p1}{a matrix with n rows and 2 columms specifying n points in 2D space.}

\item{p2}{a matrix with n rows and 2 columns specifying n points in 2D space.}
}
\value{
The distances between each corresponding pair of points.
}
\description{
This function uses linear algebra and vectorized operations to
  quickly calculate distances between many pairs of points on a 2D coordinate
  plane.
}
