% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bouts.R
\name{map_times_distances}
\alias{map_times_distances}
\title{Create Mappings Between Times and Distances}
\usage{
map_times_distances(distances, integer_times)
}
\arguments{
\item{distances}{distances traveled between consecutive points. These should
have been calculated with `linear_algebra_distances()`.}

\item{integer_times}{times as integers. These should have been calculated
with `scale_times_to_integers()`}
}
\value{
a list of two vectors that use integer times as indices (keys).
  `expanded_distances` maps integer times to distances. `integer_times_map`
  maps integer times to the original distance indices.
}
\description{
With times as integers, this function creates mappings using
  integer times as indices (keys).
}
