% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bouts.R
\name{max_decimal_places}
\alias{max_decimal_places}
\title{Find the Maximum Amount of Decimal Places}
\usage{
max_decimal_places(numbers)
}
\arguments{
\item{numbers}{a vector of numbers.}
}
\value{
the maximum amount of decimal places found in any number.
}
\description{
Given a numeric vector, find the maximum amount of decimal
  places among all elements.
}
