% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the Distribution of a Vector}
\usage{
plot_distribution(
  observations,
  type,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  breaks = "Sturges",
  xlim = range(observations)
)
}
\arguments{
\item{observations}{a vector of numeric observations.}

\item{type}{the type of plot. Supported types are `"histogram"`, `"kde"`, and
`"ecdf"`.}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x-axis.}

\item{ylab}{a label for the y-axis.}

\item{breaks}{breaks in a histogram. If the plot is not a histogram, this
argument does nothing. See the `breaks` argument in `graphics::hist()` for
more details.}

\item{xlim}{the x-limits of the plot. This may be useful for histograms.}
}
\description{
This is a simple function for plotting the distribution of a
  vector. Outputs from `calculate_angles()`,
  `calculate_bouts_per_time_unit()`, and `calculate_distances()` can be
  visualized with this function, although users may need to create their own
  plots if more advanced capabilities are required.
}
\examples{
trajectory <- create_trajectory(cadmium_zebrafish,
                                x = "X.center",
                                y = "Y.center",
                                t = "Trial.time")
bouts <- find_bouts(trajectory, min_distance = 2, min_time = 0.6)
angles <- calculate_angles(trajectory, bouts = bouts)
plot_distribution(angles, "histogram",
                  main = "Distribution of Angles",
                  xlab = "Angle (degrees)",
                  ylab = "Density",
                  breaks = 36,
                  xlim = c(-180, 180))

}
