% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_zltrajectory.R
\name{run_zltrajectory}
\alias{run_zltrajectory}
\title{Run the zltrajectory Shiny App}
\usage{
run_zltrajectory()
}
\description{
This function runs the Shiny web application for this package,
  providing an interactive interface for a common workflow provided by
  zltrajectory. This function does not return, but a webpage will be opened.
}
\examples{
\dontrun{
run_zltrajectory()
}

}
