% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bouts.R
\name{scale_times_to_integers}
\alias{scale_times_to_integers}
\title{Convert Times to Positive Integers}
\usage{
scale_times_to_integers(trajectory, min_time)
}
\arguments{
\item{trajectory}{a trajectory data frame.}

\item{min_time}{the `min_time` argument from `find_bouts()`.}
}
\value{
a list containing trajectory times with the name `integer_times` and
  `min_time`, both modified to be positive integers.
}
\description{
This function is used within `find_bouts()` to convert
  trajectory times and the `min_time` argument to positive integers by
  multiplying them with the smallest power of 10 that removes all decimal
  places in each time. If 0 is included in the trajectory times, each
  trajectory time is additionally increased by 1. Ultimately, by scaling
  trajectory times to positive integers, they can be used as vector indices.
}
