% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{validate_data}
\alias{validate_data}
\title{Validate Data Used to Create a Trajectory}
\usage{
validate_data(data, x, y, t = NULL, rate = NULL)
}
\arguments{
\item{data}{the `data` argument from `create_trajectory()`.}

\item{x}{the `x` argument from `create_trajectory()`.}

\item{y}{the `y` argument from `create_trajectory()`.}

\item{t}{the `t` argument from `create_trajectory()`.}

\item{rate}{the `rate` argument from `create_trajectory()`.}
}
\description{
This function performs validation checks on the arguments used
  in `create_trajectory()`. There is no return value.
}
