% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{validate_diagnostic_time_indices}
\alias{validate_diagnostic_time_indices}
\title{Validate Time Indices for Diagnostic Plots}
\usage{
validate_diagnostic_time_indices(time_indices, col_time_indices)
}
\arguments{
\item{time_indices}{the `time_indices` argument from `plot_diagnostics()`.}

\item{col_time_indices}{the `col_time_indices` argument from
`plot_diagonstics()`.}
}
\description{
This function performs validation checks on the `time_indices`
  and `col_time_indices` arguments in `plot_diagnostics()`. There is no
  return value.
}
