% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{validate_matrices}
\alias{validate_matrices}
\title{Validate Matrices for Linear Algebra Functions}
\usage{
validate_matrices(m1, m2)
}
\arguments{
\item{m1}{a matrix.}

\item{m2}{a matrix.}
}
\description{
This function performs validation checks on matrices used in
  `linear_algebra_angles()` and `linear_algebra_distances()`. There is no
  return value.
}
