% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{validate_time_limits}
\alias{validate_time_limits}
\title{Validate Time Limits on Plots}
\usage{
validate_time_limits(time_min, time_max)
}
\arguments{
\item{time_min}{the minimum time to plot.}

\item{time_max}{the maximum time to plot.}
}
\description{
This function performs validation checks on the minimum and
  maximum times to show in plots that display time on the x-axis. There is no
  return value.
}
