% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_args.R
\name{validate_trajectory}
\alias{validate_trajectory}
\title{Validate a Trajectory Data Frame}
\usage{
validate_trajectory(trajectory, bouts = NULL)
}
\arguments{
\item{trajectory}{a trajectory data frame.}

\item{bouts}{a vector of indices corresponding to the beginnings of bouts.}
}
\description{
This function validates a trajectory data frame created from
  `create_trajectory()`. Additionally, this function can also validate bout
  indices obtained from `find_bouts()`. There is no return value.
}
