% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot the Raw Trajectory}
\usage{
plot_trajectory(
  trajectory,
  main = "Trajectory",
  xlab = "x-position",
  ylab = "y-position",
  time_min = min(trajectory$t),
  time_max = max(trajectory$t),
  col = NULL
)
}
\arguments{
\item{trajectory}{a trajectory data frame. See `create_trajectory()` for more
details.}

\item{main}{a main title for the plot.}

\item{xlab}{a label for the x-axis.}

\item{ylab}{a label for the y-axis.}

\item{time_min}{the minimum time to plot.}

\item{time_max}{the maximum time to plot.}

\item{col}{colors of each point in the plot. By default the rainbow color
palette is used, with red indicating the position at the earliest time and
magenta indicating the position at the latest time.}
}
\description{
This function plots the x- and y-positions of a fish within a
  specified time range.
}
\examples{
trajectory <- create_trajectory(cadmium_zebrafish,
                                x = "X.center",
                                y = "Y.center",
                                t = "Trial.time")
plot_trajectory(trajectory,
                xlab = "x-position (mm)", ylab = "y-position (mm)",
                time_min = 1200, time_max = 1260)

}
